% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{basehaz.coxmtl}
\alias{basehaz.coxmtl}
\title{Predict the cumulative baseline hazard function for \code{coxmtl} objects}
\usage{
\method{basehaz}{coxmtl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxmtl}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{data.frame} with event time, cumulative baseline hazard, and
strata columns.
}
\description{
Predict the cumulative baseline hazard function for \code{coxmtl} objects
}
