% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{basehaz.coxtrans}
\alias{basehaz.coxtrans}
\title{Predict the cumulative baseline hazard function for \code{coxtrans} objects}
\usage{
\method{basehaz}{coxtrans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{data.frame} with one row for each time point, and columns
containing the event time, the cumulative baseline hazard function, and the
strata.
}
\description{
Predict the cumulative baseline hazard function for \code{coxtrans} objects
}
