% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz.R
\name{basehaz}
\alias{basehaz}
\title{Generic function for basehaz}
\usage{
basehaz(object, ...)
}
\arguments{
\item{object}{Any object.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of baseline hazard.
}
\description{
Generic function for basehaz
}
