% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.coxens.R
\name{BIC.coxens}
\alias{BIC.coxens}
\title{BIC method for a \code{coxens} model}
\usage{
\method{BIC}{coxens}(object, type = c("traditional", "modified"), ...)
}
\arguments{
\item{object}{a \code{coxens} object.}

\item{type}{A character string specifying the type of BIC to compute.
"traditional" corresponds to Cn=1, and "modified" corresponds to
Cn=log(log(d)).}

\item{...}{Unused.}
}
\value{
The BIC of the model.
}
\description{
BIC method for a \code{coxens} model
}
