% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{BIC.coxmtl}
\alias{BIC.coxmtl}
\title{Bayesian Information Criterion for \code{coxmtl} objects}
\usage{
\method{BIC}{coxmtl}(object, ...)
}
\arguments{
\item{object}{A \code{coxmtl} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Bayesian Information Criterion for \code{coxmtl} objects
}
