% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{BIC.coxtrans}
\alias{BIC.coxtrans}
\title{BIC for a \code{coxtrans} object}
\usage{
\method{BIC}{coxtrans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric scalar, the BIC value (smaller is better).
}
\description{
Computes the Bayesian Information Criterion using the full stratified
partial log-likelihood and the number of free (non-zero) parameters
returned by \code{\link{coef.coxtrans}}.
}
