% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{BIC.ncvcox}
\alias{BIC.ncvcox}
\title{BIC for a \code{ncvcox} object}
\usage{
\method{BIC}{ncvcox}(object, type = c("traditional", "modified"), ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{type}{A character string specifying the type of BIC to compute.
"traditional" corresponds to Cn=1, and "modified" corresponds to
Cn=log(log(d)).}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric value representing the BIC of the fitted \code{ncvcox}
object.
}
\description{
BIC for a \code{ncvcox} object
}
