% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.coxens.R
\name{coef.coxens}
\alias{coef.coxens}
\title{Extract Model Coefficients from \code{coxens} Object}
\usage{
\method{coef}{coxens}(object, ...)
}
\arguments{
\item{object}{the result of a \code{coxens} fit.}

\item{...}{Unused.}
}
\value{
a vector of coefficients.
}
\description{
Extract Model Coefficients from \code{coxens} Object
}
