% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{coef.coxmtl}
\alias{coef.coxmtl}
\title{Extract the coefficients from a \code{coxmtl} object}
\usage{
\method{coef}{coxmtl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxmtl}.}

\item{...}{Additional arguments (unused).}
}
\value{
A named numeric vector containing the free parameters of the fitted
\code{coxmtl} object.
}
\description{
Extract the coefficients from a \code{coxmtl} object
}
