% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{coef.coxtrans}
\alias{coef.coxtrans}
\title{Extract the coefficients from a \code{coxtrans} object}
\usage{
\method{coef}{coxtrans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments (unused).}
}
\value{
A named numeric vector containing the coefficients of the fitted
\code{coxtrans} object. The names indicate the group(s) to which the
coefficients belong. Zero coefficients are removed.
}
\description{
Extract the coefficients from a \code{coxtrans} object
}
