% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv_ncvcox.R
\name{coef.cv_ncvcox}
\alias{coef.cv_ncvcox}
\title{Extract coefficients from a cv_ncvcox object}
\usage{
\method{coef}{cv_ncvcox}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{a cv_ncvcox object.}

\item{s}{a character string specifying the rule to select the coefficients.
The default is "lambda.min". Other options are "lambda.1se".}

\item{...}{Unused.}
}
\value{
a vector of coefficients.
}
\description{
Extract coefficients from a cv_ncvcox object
}
