% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{coef.ncvcox}
\alias{coef.ncvcox}
\title{Extract the coefficients from a \code{ncvcox} object}
\usage{
\method{coef}{ncvcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{...}{Additional arguments (unused).}
}
\value{
A named numeric vector containing the coefficients of the fitted
\code{ncvcox} object. Zero coefficients are removed.
}
\description{
Extract the coefficients from a \code{ncvcox} object
}
