% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxens.R
\name{coxens}
\alias{coxens}
\title{Multiple-sources Cox proportional hazards model with group sparsity
and Local-Global transfer}
\usage{
coxens(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{lambda1}{a non-negative value specifying the sparse penalty
parameter. The default is 0.}

\item{lambda2}{a non-negative value specifying the global biased penalty
parameter. The default is 0.}

\item{lambda3}{a non-negative value specifying the local biased penalty
parameter. The default is 0.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{rho}{a value in (2, 10) specifying the expansion factor of the
augmented Lagrangian's penalty parameter. The default is 2.0.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a coxsg object.
}
\description{
Multiple-sources Cox proportional hazards model with group sparsity
and Local-Global transfer
}
\examples{
library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim2$group)
fit <- coxens(
  formula, sim2, group,
  lambda1 = 2e-2, lambda3 = 3e-3, penalty = "SCAD"
)
summary(fit)
}
