% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl_oracle.R
\name{coxmtl_oracle}
\alias{coxmtl_oracle}
\title{Multi-task Learning of Cox proportional hazards model (Oracle)}
\usage{
coxmtl_oracle(
  formula,
  data,
  group,
  sparse_idx,
  group_idx,
  rho = 2,
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{sparse_idx}{a logical matrix specifying the sparse parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{group_idx}{a logical matrix specifying the group parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{rho}{a value in (2, 10) specifying the expansion factor of the
augmented Lagrangian's penalty parameter. The default is 2.0.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a coxmtl object.
}
\description{
Multi-task Learning of Cox proportional hazards model (Oracle)
}
