% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxsg_oracle.R
\name{coxsg_oracle}
\alias{coxsg_oracle}
\title{Personalized Multi-task Learning of Cox proportional hazards model (Oracle)}
\usage{
coxsg_oracle(formula, data, group, sparse_idx, group_idx, init, control, ...)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{sparse_idx}{a logical matrix specifying the sparse parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{group_idx}{a logical matrix specifying the group parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a coxmtl object.
}
\description{
Personalized Multi-task Learning of Cox proportional hazards model (Oracle)
}
\examples{
formula <- Surv(time, status) ~ . - group - id
sparse_idx <- matrix(
  TRUE, nrow = 10, ncol = 10
)
group_idx <- matrix(
  rep(c(rep(1, 10), rep(c(1, 2), 5)), 5),
  ncol = 10, byrow = TRUE
)
fit <- coxsg_oracle(
  formula, sim2, as.factor(sim2$group), sparse_idx, group_idx
)
fit$eta
}
