% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ncvcox.R
\name{cv.ncvcox}
\alias{cv.ncvcox}
\title{Cross-validation for \code{ncvcox}}
\usage{
cv.ncvcox(
  formula,
  data,
  group,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a \code{~}
operator, and the terms on the right. The response must be a survival object
as returned by the Surv function.}

\item{data}{A data frame containing the variables in the model.}

\item{group}{A factor variable indicating the group of each observation.}

\item{penalty}{A character string specifying the penalty function. The
default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{A non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{nfolds}{An integer specifying the number of folds.}

\item{nlambdas}{An integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{An integer specifying the random seed. Default is 0.}

\item{control}{An object of class \link{survtrans_control} containing control
parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Additional arguments to be passed to the fitting algorithm.}
}
\value{
An object of class \code{cv.ncvcox}.
}
\description{
Cross-validation for \code{ncvcox}
}
\examples{
formula <- survival::Surv(time, status) ~ . - group - id
df <- sim2[sim2$group == 2 | sim2$group == 4, ]
cvfit <- cv.ncvcox(formula, df, df$group, penalty = "SCAD")
coef(cvfit)
}
