% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_coxtrans.R
\name{cv_coxtrans}
\alias{cv_coxtrans}
\title{Cross-validation for \code{coxtrans}}
\usage{
cv_coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{fit_source}{a coxph object fitted to the source data.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 1.5 for MCP.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{an integer specifying the random seed.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a cv_coxtrans object.
}
\description{
Cross-validation for \code{coxtrans}
}
\examples{
library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
cv_coxtrans(sim1_trg, fit_src)
}
