% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ncvcox.R
\name{cv_ncvcox}
\alias{cv_ncvcox}
\title{Cross-validation for \code{coxtrans}}
\usage{
cv_ncvcox(
  formula,
  data,
  offset,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms, separated by + operators, on the right.}

\item{data}{a data frame containing the variables in the model.}

\item{offset}{a numeric vector specifying the offset.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3 for MCP.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{an integer specifying the random seed.}

\item{control}{Object of class \link{survtrans.control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans.control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans.control}}.}
}
\value{
a cv_ncvcox object.
}
\description{
Cross-validation for \code{coxtrans}
}
