% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_odetrans.R
\name{cv_odetrans}
\alias{cv_odetrans}
\title{Cross-validation for \code{survtran}}
\usage{
cv_odetrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  cbh_func = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{fit_source}{a survode object.}

\item{penalty}{a character string specifying the type of penalty.
Possible values are \code{"lasso"}, \code{"MCP"} and \code{"SCAD"}.}

\item{gamma}{a numeric value specifying the tuning parameter for MCP or SCAD.}

\item{cbh_func}{a function specifying the cumulative baseline hazard. The
default is NULL, which means the cumulative baseline hazard is estimated
from the survode object fitted to the source data.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{an integer specifying the random seed.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a cv_survtran object.
}
\description{
Cross-validation for \code{survtran}
}
\examples{
formula <- Surv(time, status) ~ . - group - id
fit_src <- survode(formula, data = sim1[sim1$group == 1, ], df = 10)
cv_fit <- cv_odetrans(sim1[sim1$group == 2, ], fit_src, lambda = 0.1)
}
