% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_survtrans.R
\name{cv_survtrans}
\alias{cv_survtrans}
\title{Cross-validation for \code{survtran}}
\usage{
cv_survtrans(
  data,
  fit_source,
  penalty = "lasso",
  gamma = NULL,
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{fit_source}{a survode object.}

\item{penalty}{a character string specifying the type of penalty.
Possible values are \code{"lasso"}, \code{"MCP"} and \code{"SCAD"}.}

\item{gamma}{a numeric value specifying the tuning parameter for MCP or SCAD.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{an integer specifying the random seed.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a cv_survtran object.
}
\description{
Cross-validation for \code{survtran}
}
\examples{
library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
cv_survtrans(sim1_trg, fit_src)
}
