% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagnose}
\alias{diagnose}
\title{Generic function for diagnose}
\usage{
diagnose(object, ...)
}
\arguments{
\item{object}{Any object.}

\item{...}{Additional arguments.}
}
\value{
Called for its side effect of producing diagnostic plots.
Returns \code{NULL} invisibly.
}
\description{
Generic function for diagnose
}
