% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxens.R
\name{logLik.coxens}
\alias{logLik.coxens}
\title{BIC method for a \code{coxens} model}
\usage{
\method{logLik}{coxens}(object, ...)
}
\arguments{
\item{object}{the result of a \code{coxens} fit.}

\item{...}{Unused.}
}
\value{
The log-likelihood of the model.
}
\description{
BIC method for a \code{coxens} model
}
