% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{logLik.coxmtl}
\alias{logLik.coxmtl}
\title{Log-likelihood for a \code{coxmtl} object}
\usage{
\method{logLik}{coxmtl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxmtl}.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric value representing the log-likelihood of the fitted
\code{coxmtl} object.
}
\description{
Log-likelihood for a \code{coxmtl} object
}
