% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxsg.R
\name{logLik.coxsg}
\alias{logLik.coxsg}
\title{Log-likelihood of a subgroup analysis for Cox proportional hazards model}
\usage{
\method{logLik}{coxsg}(object, data, group, ...)
}
\arguments{
\item{object}{a coxsg object.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{...}{Unused.}
}
\value{
the log-likelihood of the model.
}
\description{
Log-likelihood of a subgroup analysis for Cox proportional hazards model
}
