% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxtl.R
\name{logLik.coxtl}
\alias{logLik.coxtl}
\title{Log-likelihood of Transfer Learning for Cox proportional hazards model}
\usage{
\method{logLik}{coxtl}(object, data, group, ...)
}
\arguments{
\item{object}{a coxtl object.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{...}{Unused.}
}
\value{
the log-likelihood of the model.
}
\description{
Log-likelihood of Transfer Learning for Cox proportional hazards model
}
