% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{logLik.coxtrans}
\alias{logLik.coxtrans}
\title{Log-likelihood for a \code{coxtrans} object}
\usage{
\method{logLik}{coxtrans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric value representing the log-likelihood of the fitted
\code{coxtrans} object.
}
\description{
Log-likelihood for a \code{coxtrans} object
}
