% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{logLik.ncvcox}
\alias{logLik.ncvcox}
\title{Log-likelihood for a \code{ncvcox} object}
\usage{
\method{logLik}{ncvcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric value representing the log-likelihood of the fitted
\code{ncvcox} object.
}
\description{
Log-likelihood for a \code{ncvcox} object
}
