% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{ncvcox}
\alias{ncvcox}
\title{Non-convex penalized Cox proportional hazards model}
\usage{
ncvcox(
  formula,
  data,
  group,
  lambda = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a \code{~}
operator, and the terms on the right. The response must be a survival object
as returned by the Surv function.}

\item{data}{A data frame containing the variables in the model.}

\item{group}{A factor variable indicating the group of each observation.}

\item{lambda}{A non-negative value specifying the penalty parameter. The
default is 0.}

\item{penalty}{A character string specifying the penalty function. The
default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{A non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{init}{A numeric vector of initial values for the coefficients. The
default is a zero vector.}

\item{control}{An object of class \link{survtrans_control} containing control
parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Additional arguments to be passed to the fitting algorithm.}
}
\value{
An object of class \code{ncvcox}.
}
\description{
Non-convex penalized Cox proportional hazards model
}
\examples{
formula <- Surv(time, status) ~ . - group - id
df <- sim2[sim2$group == 2 | sim2$group == 4, ]
fit <- ncvcox(formula, df, df$group, lambda = 0.1, penalty = "SCAD")
summary(fit)
}
