% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odetrans.R
\name{odetrans}
\alias{odetrans}
\title{Transfer Learning for Survival Analysis via Ordinary Differential Equations}
\usage{
odetrans(
  data,
  fit_source,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  lambda = 0,
  cbh_func = NULL,
  init,
  control,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{fit_source}{a survode object fitted to the source data.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{lambda}{a non-negative value specifying the penalty parameter. The
default is 0.}

\item{cbh_func}{a function specifying the cumulative baseline hazard. The
default is NULL, which means the cumulative baseline hazard is estimateds
from the survode object fitted to the source data.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\description{
Transfer Learning for Survival Analysis via Ordinary Differential Equations
}
\examples{
library(survode)
library(survtrans)
fit_src <- survode(Surv(time, status) ~ ., data = sim1_src, df = 10)
odetrans(sim1_trg, fit_src, lambda = 0.1)
}
