% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.coxtrans.R
\name{plot.cv.coxtrans}
\alias{plot.cv.coxtrans}
\title{Plot cross-validation curve for a \code{cv.coxtrans} object}
\usage{
\method{plot}{cv.coxtrans}(x, ...)
}
\arguments{
\item{x}{A \code{cv.coxtrans} object.}

\item{...}{Further graphical arguments passed to \code{plot}.}
}
\description{
Plots the CV deviance profile along \code{lambda1} with \code{lambda2} and
\code{lambda3} fixed at their optimal values, in the style of
\code{plot.cv.glmnet}.
}
