% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.ncvcox.R
\name{plot.cv.ncvcox}
\alias{plot.cv.ncvcox}
\title{Plot the cross-validation curve from \code{cv.ncvcox} object}
\usage{
\method{plot}{cv.ncvcox}(x, ...)
}
\arguments{
\item{x}{a cv.ncvcox object.}

\item{...}{Unused.}
}
\description{
Plot the cross-validation curve from \code{cv.ncvcox} object
}
