% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv_ncvcox.R
\name{plot.cv_ncvcox}
\alias{plot.cv_ncvcox}
\title{Plot the cross-validation curve from a cv_ncvcox x}
\usage{
\method{plot}{cv_ncvcox}(x, ...)
}
\arguments{
\item{x}{a cv_ncvcox x.}

\item{...}{Unused.}
}
\description{
Plot the cross-validation curve from a cv_ncvcox x
}
