% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{print.coxmtl}
\alias{print.coxmtl}
\title{Print method for a \code{coxmtl} object}
\usage{
\method{print}{coxmtl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{coxmtl}.}

\item{...}{Additional arguments passed to \code{print.summary.coxmtl}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for a \code{coxmtl} object
}
