% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{print.coxtrans}
\alias{print.coxtrans}
\title{Print method for a \code{coxtrans} object}
\usage{
\method{print}{coxtrans}(x, ...)
}
\arguments{
\item{x}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments passed to \code{print.summary.coxtrans}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for a \code{coxtrans} object
}
