% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{print.refit.coxmtl}
\alias{print.refit.coxmtl}
\title{Print method for a \code{refit.coxmtl} object}
\usage{
\method{print}{refit.coxmtl}(
  x,
  digits = max(getOption("digits") - 3, 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A \code{refit.coxmtl} object.}

\item{digits}{Number of significant digits.}

\item{signif.stars}{Logical; print significance stars.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for a \code{refit.coxmtl} object
}
