% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1_src.R
\docType{data}
\name{sim1_src}
\alias{sim1_src}
\title{Simulated Source Dataset for Example 1}
\format{
A data frame with 1000 rows and 12 variables:
\describe{
\item{time}{survival time}
\item{status}{censoring status}
\item{X1}{covariate 1}
\item{X2}{covariate 2}
\item{X3}{covariate 3}
\item{X4}{covariate 4}
\item{X5}{covariate 5}
\item{X6}{covariate 6}
\item{X7}{covariate 7}
\item{X8}{covariate 8}
\item{X9}{covariate 9}
\item{X10}{covariate 10}
}
}
\description{
Simulated Source Dataset for Example 1
}
