% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simsurv_tl}
\alias{simsurv_tl}
\title{Simulate survival data for a multi-source Cox model}
\usage{
simsurv_tl(
  beta,
  eta,
  lambda,
  gamma,
  dist,
  maxt,
  n_samples,
  seed = 0,
  sigma = NULL
)
}
\arguments{
\item{beta}{A vector of length p representing the common coefficients,
where p is the number of features.}

\item{eta}{A matrix of size p x K representing the group-specific
coefficients, where K is the number of groups.}

\item{lambda}{A vector of length K representing the baseline hazard's scale
parameters.}

\item{gamma}{A vector of length K representing the baseline hazard's shape
parameters.}

\item{dist}{A string specifying the distribution of the baseline hazard,
either "exponential", "weibull", or "gompertz".}

\item{maxt}{A positive number specifying the maximum time to simulate.}

\item{n_samples}{A vector of length K specifying the number of samples per
group, or a single number specifying the total number of samples.}

\item{seed}{An integer specifying the random seed, with a default value of 0.}

\item{sigma}{An optional p x p covariance matrix for the covariates. If
\code{NULL} (default), an identity matrix is used.}
}
\value{
A data frame with columns "id", "group", "X1", "X2", ..., "Xp",
"time", and "status".
}
\description{
Simulate survival data for a multi-source Cox model
}
\examples{
beta <- c(1, 1)
eta <- matrix(c(0, 0, 1, 1), nrow = 2, ncol = 2)
lambda <- c(1, 2)
gamma <- c(2, 1)
dist <- c("gompertz", "weibull")
maxt <- 3
n_samples <- 100
df <- simsurv_tl(beta, eta, lambda, gamma, dist, maxt, n_samples)
df
}
