% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{summary.ncvcox}
\alias{summary.ncvcox}
\title{Summary method for a \code{ncvcox} object}
\usage{
\method{summary}{ncvcox}(object, conf.int = 0.95, compressed = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{conf.int}{A numeric value between 0 and 1 indicating the confidence
level of the confidence interval. Default is 0.95.}

\item{compressed}{Logical; if \code{TRUE}, the summary is compressed and
only includes the group-level coefficients. Default is \code{TRUE}.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{summary.ncvcox}, with the following
components:
\item{\code{n}, \code{nevent}}{Number of observations and number of events,
respectively, in the fit.}
\item{\code{logLik}}{The log partial likelihood at the final value.}
\item{\code{coefficients}}{A matrix with one row for each coefficient, and
columns containing the coefficient, the hazard ratio exp(coef), standard
error, Wald statistic, and P value.}
\item{\code{conf.int}}{A matrix with one row for each coefficient, containing
the confidence limits for exp(coef).}
}
\description{
Summary method for a \code{ncvcox} object
}
