% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtrans-package.R
\docType{package}
\name{survtrans-package}
\alias{survtrans}
\alias{survtrans-package}
\title{survtrans: Transfer Learning for Survival Analysis}
\description{
Transfer learning for survival analysis. This package provides a framework for transferring survival information from source domain(s) to target domain. The package now only supports the Cox proportional hazards model with global and local transfer learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ziyangg98/survtrans}
  \item \url{http://gongziyang.com/survtrans/}
  \item Report bugs at \url{https://github.com/ziyangg98/survtrans/issues}
}

}
\author{
\strong{Maintainer}: Ziyang Gong \email{gongziyang@smail.swufe.edu.cn} (\href{https://orcid.org/0009-0001-4870-9102}{ORCID})

}
\keyword{internal}
