% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtrans.control.R
\name{survtrans.control}
\alias{survtrans.control}
\title{Ancillary arguments for controlling survtrans fitting}
\usage{
survtrans.control(eps = 1e-06, maxit = 1000)
}
\arguments{
\item{eps}{the convergence criteria for the CCD algorithm. Default is 1e-6.}

\item{maxit}{the maximum number of iterations for the CCD algorithm.
Default is 1000.}
}
\description{
Ancillary arguments for controlling survtrans fitting
}
