% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.coxens.R
\name{vcov.coxens}
\alias{vcov.coxens}
\title{Variance-covariance method for a \code{coxens} model}
\usage{
\method{vcov}{coxens}(object, ...)
}
\arguments{
\item{object}{the result of a \code{coxens} fit.}

\item{...}{Unused.}
}
\value{
A matrix representing the variance-covariance matrix of
the coefficients.
}
\description{
Variance-covariance method for a \code{coxens} model
}
