% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmtl.R
\name{vcov.coxmtl}
\alias{vcov.coxmtl}
\title{Variance-covariance matrix for a \code{coxmtl} object}
\usage{
\method{vcov}{coxmtl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxmtl}.}

\item{...}{Additional arguments (unused).}
}
\value{
A matrix representing the variance-covariance matrix of the free
parameters.
}
\description{
Variance-covariance matrix for a \code{coxmtl} object
}
