% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{vcov.coxtrans}
\alias{vcov.coxtrans}
\title{Variance-covariance matrix for a \code{coxtrans} object.}
\usage{
\method{vcov}{coxtrans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{...}{Additional arguments (unused).}
}
\value{
A matrix representing the variance-covariance matrix of the
coefficients.
}
\description{
Variance-covariance matrix for a \code{coxtrans} object.
}
