% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.ncvcox.R
\name{vcov.ncvcox}
\alias{vcov.ncvcox}
\title{Variance-covariance method for a \code{ncvcox} model}
\usage{
\method{vcov}{ncvcox}(object, ...)
}
\arguments{
\item{object}{the result of a \code{ncvcox} fit.}

\item{...}{Unused.}
}
\value{
A matrix representing the variance-covariance matrix of
the coefficients.
}
\description{
Variance-covariance method for a \code{ncvcox} model
}
