% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{basehaz.ncvcox}
\alias{basehaz.ncvcox}
\title{Predict the cumulative baseline hazard function for \code{ncvcox} objects}
\usage{
\method{basehaz}{ncvcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{data.frame} with one row for each time point, and columns
containing the event time, the cumulative baseline hazard function, and the
strata.
}
\description{
Predict the cumulative baseline hazard function for \code{ncvcox} objects
}
