% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.coxmtl.R
\name{BIC.coxmtl}
\alias{BIC.coxmtl}
\title{BIC for Multi-source Cox proportional hazards model}
\usage{
\method{BIC}{coxmtl}(object, data, group, ...)
}
\arguments{
\item{object}{a coxmtl object.}

\item{data}{a data frame containing the variables in the model, which
should be the same as the data used to fit the model.}

\item{group}{a factor specifying the group of each sample, which
should be the same as the group used to fit the model.}

\item{...}{Unused.}
}
\value{
the BIC of the model.
}
\description{
BIC for Multi-source Cox proportional hazards model
}
