% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{BIC.coxtrans}
\alias{BIC.coxtrans}
\title{BIC for a \code{coxtrans} object}
\usage{
\method{BIC}{coxtrans}(object, type = c("traditional", "modified"), ...)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{type}{A character string specifying the type of BIC to compute.
"traditional" corresponds to Cn=1, and "modified" corresponds to
Cn=log(log(d)).}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric value representing the BIC of the fitted \code{coxtrans}
object.
}
\description{
BIC for a \code{coxtrans} object
}
