% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.ncvcox.R
\name{BIC.ncvcox}
\alias{BIC.ncvcox}
\title{BIC for a \code{ncvcox} model}
\usage{
\method{BIC}{ncvcox}(object, type = c("traditional", "modified"), ...)
}
\arguments{
\item{object}{a \code{ncvcox} object.}

\item{type}{A character string specifying the type of BIC to compute.
"traditional" corresponds to Cn=1, and "modified" corresponds to
Cn=log(log(d)).}

\item{...}{Unused.}
}
\value{
the BIC of the model.
}
\description{
BIC for a \code{ncvcox} model
}
