% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lambda_max.R
\name{calc_lambda_max}
\alias{calc_lambda_max}
\title{Determine the maximum value of the penalty parameter lambda}
\usage{
calc_lambda_max(formula, data, group, offset)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{offset}{a numeric vector specifying the offset.}
}
\value{
the maximum value of the penalty parameter lambda.
}
\description{
Determine the maximum value of the penalty parameter lambda
}
