% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lambda_max.R
\name{calc_lambda_max}
\alias{calc_lambda_max}
\title{Calculate the maximum value of the penalty parameter lambda}
\usage{
calc_lambda_max(formula, data, group, offset)
}
\arguments{
\item{formula}{A formula expression for regression models, in the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{A data frame containing the variables in the model.}

\item{group}{A factor specifying the group of each sample.}

\item{offset}{A numeric vector specifying the offset.}
}
\value{
The maximum value of the penalty parameter lambda, which shrinks all
the coefficients to zero.
}
\description{
Calculate the maximum value of the penalty parameter lambda
}
