% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv.ncvcox.R
\name{coef.cv.ncvcox}
\alias{coef.cv.ncvcox}
\title{Extract coefficients from a \code{cv.ncvcox} object}
\usage{
\method{coef}{cv.ncvcox}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{An object of class \code{cv.ncvcox}.}

\item{s}{A character string specifying the rule to select the coefficients.
The default is "lambda.min". Other options are "lambda.1se".}

\item{...}{Unused.}
}
\value{
A named numeric vector containing the coefficients of the fitted
\code{cv.ncvcox} object.
}
\description{
Extract coefficients from a \code{cv.ncvcox} object
}
