% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ncvcox.R
\name{coef.ncvcox}
\alias{coef.ncvcox}
\title{Extract coefficients from a ncvcox object}
\usage{
\method{coef}{ncvcox}(object, ...)
}
\arguments{
\item{object}{a ncvcox object.}

\item{...}{Unused.}
}
\value{
a vector of coefficients.
}
\description{
Extract coefficients from a ncvcox object
}
