% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtl_oracle.R
\name{coxtl_oracle}
\alias{coxtl_oracle}
\title{Transfer Learning for Cox proportional hazards model (Oracle)
without constraints}
\usage{
coxtl_oracle(
  formula,
  data,
  group,
  target,
  sparse_idx,
  group_idx,
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{target}{a factor specifying the target group.}

\item{sparse_idx}{a logical matrix specifying the sparse parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{group_idx}{a logical matrix specifying the group parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a coxtl_oracle object.
}
\description{
Transfer Learning for Cox proportional hazards model (Oracle)
without constraints
}
\examples{
formula <- Surv(time, status) ~ . - group - id
sparse_idx <- matrix(FALSE, nrow = 10, ncol = 9)
for (i in 1:9) {
  if (i \%\% 2 == 1) sparse_idx[, i] <- rep(c(FALSE, TRUE), 5)
}
group_idx <- matrix(
  rep(c(rep(1, 10), rep(c(1, 2), 5)), 5),
  ncol = 10, byrow = TRUE
)
group_idx <- group_idx[, -1]

# Sparse
fit <- coxtl_oracle(
  formula, sim2, as.factor(sim2$group),
  target = 10, sparse_idx = sparse_idx
)
fit$beta
fit$eta

# Group
fit <- coxtl_oracle(
  formula, sim2, as.factor(sim2$group),
  target = 10, group_idx = group_idx
)
fit$beta
fit$eta

# Sparse and group
fit <- coxtl_oracle(
  formula, sim2, as.factor(sim2$group),
  target = 10, sparse_idx = sparse_idx, group_idx = group_idx
)
fit$beta
fit$eta
}
