% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{coxtrans}
\alias{coxtrans}
\title{Transfer Learning for Cox Proportional Hazards Model}
\usage{
coxtrans(
  data,
  fit_source,
  penalty = "lasso",
  lambda = 0,
  gamma = NULL,
  init,
  control,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{fit_source}{a coxph object fitted to the source data.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{lambda}{a non-negative value specifying the penalty parameter. The
default is 0.}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 1.5 for MCP.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\description{
Transfer Learning for Cox Proportional Hazards Model
}
\examples{
library(survival)
library(survtrans)
fit_src <- coxph(Surv(time, status) ~ ., data = sim1_src)
coxtrans(sim1_trg, fit_src, lambda = 0.1)
}
