% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{coxtrans}
\alias{coxtrans}
\title{Transfer Learning for Cox Model with Global and Local Shrinkage}
\usage{
coxtrans(
  formula,
  data,
  group,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  rho = 2,
  tau = 10,
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a \code{~}
operator, and the terms on the right. The response must be a survival object
as returned by the Surv function.}

\item{data}{A data frame containing the variables in the model.}

\item{group}{A factor variable indicating the group of each observation.}

\item{lambda1}{A non-negative value specifying the sparse penalty parameter.
The default is 0.}

\item{lambda2}{A non-negative value specifying the global bias penalty
parameter. The default is 0.}

\item{lambda3}{A non-negative value specifying the local bias penalty
parameter. The default is 0.}

\item{penalty}{A character string specifying the penalty function. The
default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{A non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{rho}{A value larger than 1 specifying the increase/decrease factor
for the augmented Lagrangian's penalty parameter. The default is 2.0.}

\item{tau}{A value larger than 1 specifying the tolerance for the
trade-off between the primal and dual residuals. The default is 10.0.}

\item{init}{A numeric vector of initial values for the coefficients. The
default is a zero vector.}

\item{control}{An object of class \link{survtrans_control} containing control
parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Additional arguments to be passed to the fitting algorithm.}
}
\value{
An object of class \code{coxtrans}.
}
\description{
Transfer Learning for Cox Model with Global and Local Shrinkage
}
\examples{
formula <- survival::Surv(time, status) ~ . - group - id
fit <- coxtrans(
  formula, sim2, sim2$group,
  lambda1 = 0.03, lambda2 = 0.01, lambda3 = 0.01, penalty = "SCAD"
)
summary(fit)
}
