% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans.R
\name{coxtrans}
\alias{coxtrans}
\title{Transfer Learning for Cox Model with Global and Local Shrinkage}
\usage{
coxtrans(
  formula,
  data,
  group,
  target,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  vartheta = 1,
  control,
  ...
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a \code{~}
operator, and the terms on the right. The response must be a survival object
as returned by the Surv function.}

\item{data}{A data frame containing the variables in the model.}

\item{group}{A factor variable indicating the group of each observation.}

\item{target}{A character string specifying the target group.}

\item{lambda1}{A non-negative value specifying the sparse penalty parameter.
The default is 0.}

\item{lambda2}{A non-negative value specifying the global biased penalty
parameter. The default is 0.}

\item{lambda3}{A non-negative value specifying the local biased penalty
parameter. The default is 0.}

\item{penalty}{A character string specifying the penalty function. The
default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{A non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{vartheta}{A positive value specifying the fixed penalty parameter in
the augmented Lagrangian. Following Wang, Yin & Zeng (2019), this is kept
constant (not adaptive) to guarantee convergence with non-convex penalties.
The default is 1.0.}

\item{control}{An object of class \link{survtrans_control} containing control
parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Additional arguments to be passed to the fitting algorithm.}
}
\value{
An object of class \code{coxtrans}.
}
\description{
Transfer Learning for Cox Model with Global and Local Shrinkage
}
\examples{
formula <- Surv(time, status) ~ . - group - id
fit <- coxtrans(
  formula, sim2, sim2$group, 1,
  lambda1 = 0.075, lambda2 = 0.04, lambda3 = 0.04, penalty = "SCAD"
)
summary(fit)
}
