% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtrans_oracle.R
\name{coxtrans_oracle}
\alias{coxtrans_oracle}
\title{Transfer Learning of Multi-source for Cox proportional hazards model (Oracle)}
\usage{
coxtrans_oracle(formula, data, group, sparse_idx, rho = 2, init, control, ...)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{sparse_idx}{a logical matrix specifying the sparse parameters, where
each row corresponds to a feature and each column corresponds to a group.}

\item{rho}{a value in (2, 10) specifying the expansion factor of the
augmented Lagrangian's penalty parameter. The default is 2.0.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a coxtrans object.
}
\description{
Transfer Learning of Multi-source for Cox proportional hazards model (Oracle)
}
\examples{
formula <- Surv(time, status) ~ . - group - id
sparse_idx <- matrix(
  rep(rep(c(FALSE, TRUE), each = 10), 5), ncol = 10, byrow = TRUE
)
fit <- coxtrans_oracle(formula, sim2, as.factor(sim2$group), sparse_idx)
fit$eta
}
