% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_coxtrans.R
\name{cv_coxtrans}
\alias{cv_coxtrans}
\title{Cross-validation for \code{coxtrans}}
\usage{
cv_coxtrans(
  formula,
  data,
  group,
  bgroup,
  egroup,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  verbose = FALSE,
  seed = 0,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{bgroup}{a factor specifying the group of each sample for baseline
hazard.}

\item{egroup}{a factor specifying the group of each sample for effect.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{verbose}{a logical value specifying whether to print the progress.}

\item{seed}{an integer specifying the random seed.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a cv_coxtrans object.
}
\description{
Cross-validation for \code{coxtrans}
}
\examples{
library(survtrans)
formula <- Surv(time, status) ~ . - group - id
group <- as.factor(sim1$group)
cv_fit <- cv_coxtrans(formula, sim1, group, penalty = "SCAD")
coef(cv_fit, "lambda.min")
}
