% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ncvcox.R
\name{cv_ncvcox}
\alias{cv_ncvcox}
\title{Cross-validation for \code{ncvcox}}
\usage{
cv_ncvcox(
  formula,
  data,
  group,
  offset,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  nfolds = 10,
  nlambdas = 100,
  lambda_min_ratio = NULL,
  seed = 0,
  parallel = 0,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{offset}{a numeric vector specifying the offset..}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD".}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3 for MCP.}

\item{nfolds}{an integer specifying the number of folds.}

\item{nlambdas}{an integer specifying the number of lambda values.}

\item{lambda_min_ratio}{a numeric value specifying the minimum lambda value
as a fraction of lambda_max.}

\item{seed}{an integer specifying the random seed. Default is 0.}

\item{parallel}{a integer specifying the number of parallel workers.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a cv_ncvcox object.
}
\description{
Cross-validation for \code{ncvcox}
}
\examples{
library(survtrans)
formula <- Surv(time, status) ~ . - group - id
df <- sim1[sim1$group == 1, ]
cv_fit <- cv_ncvcox(formula, df, penalty = "SCAD")
coef(cv_fit)
plot(cv_fit)
}
