% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxens.R
\name{logLik.coxens}
\alias{logLik.coxens}
\title{Log-likelihood for \code{coxens} model}
\usage{
\method{logLik}{coxens}(object, ...)
}
\arguments{
\item{object}{a \code{coxens} object.}

\item{...}{Unused.}
}
\value{
The log-likelihood of the model.
}
\description{
Log-likelihood for \code{coxens} model
}
