% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxmtl.R
\name{logLik.coxmtl}
\alias{logLik.coxmtl}
\title{Log-likelihood of a Multi-task Learning for Cox proportional hazards model}
\usage{
\method{logLik}{coxmtl}(object, data, group, ...)
}
\arguments{
\item{object}{a coxmtl object.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{...}{Unused.}
}
\value{
the log-likelihood of the model.
}
\description{
Log-likelihood of a Multi-task Learning for Cox proportional hazards model
}
