% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.coxtrans.R
\name{logLik.coxtrans}
\alias{logLik.coxtrans}
\title{Log-likelihood of a Multi-source Cox proportional hazards model}
\usage{
\method{logLik}{coxtrans}(object, data, group, ...)
}
\arguments{
\item{object}{a coxtrans object.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{...}{Unused.}
}
\value{
the log-likelihood of the model.
}
\description{
Log-likelihood of a Multi-source Cox proportional hazards model
}
