% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.ncvcox.R
\name{logLik.ncvcox}
\alias{logLik.ncvcox}
\title{Log-likelihood for \code{ncvcox} model}
\usage{
\method{logLik}{ncvcox}(object, ...)
}
\arguments{
\item{object}{a \code{ncvcox} object.}

\item{...}{Unused.}
}
\value{
The log-likelihood of the model.
}
\description{
Log-likelihood for \code{ncvcox} model
}
