% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.ncvcox.R
\name{logLik.ncvcox}
\alias{logLik.ncvcox}
\title{Log-likelihood of Non-convex penalized Cox proportional hazards model}
\usage{
\method{logLik}{ncvcox}(object, ...)
}
\arguments{
\item{object}{a ncvcox object.}

\item{...}{Unused.}
}
\description{
Log-likelihood of Non-convex penalized Cox proportional hazards model
}
