% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{ncvcox}
\alias{ncvcox}
\title{Non-convex penalized Cox proportional hazards model}
\usage{
ncvcox(
  formula,
  data,
  group,
  offset,
  lambda = 0,
  penalty = c("lasso", "MCP", "SCAD"),
  gamma = switch(penalty, SCAD = 3.7, MCP = 3, 1),
  init,
  control,
  ...
)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{a factor specifying the group of each sample.}

\item{offset}{a numeric vector specifying the offset..}

\item{lambda}{a non-negative value specifying the penalty parameter. The
default is 0.}

\item{penalty}{a character string specifying the penalty function.
The default is "lasso". Other options are "MCP" and "SCAD"}

\item{gamma}{a non-negative value specifying the penalty parameter. The
default is 3.7 for SCAD and 3.0 for MCP.}

\item{init}{a numeric vector specifying the initial value of the
coefficients. The default is a zero vector.}

\item{control}{Object of class \link{survtrans_control} containing
control parameters for the fitting algorithm. Default is
\code{survtrans_control(...)}.}

\item{...}{Other arguments passed to \code{\link{survtrans_control}}.}
}
\value{
a ncvcox object.
}
\description{
Non-convex penalized Cox proportional hazards model
}
\examples{
library(survtrans)
formula <- Surv(time, status) ~ . - group - id
df_src <- sim1[sim1$group == 1, ]
df_trg <- sim1[sim1$group == 2, ]
group <- as.factor(sim1$group)
fit_src <- ncvcox(formula, df_src, lambda = 0.1, penalty = "SCAD")
offset_trg <- predict(fit_src, newdata = df_trg, type = "lp")
fit_trg <- ncvcox(
  formula, df_trg,
  offset = offset_trg, lambda = 0.2, penalty = "SCAD"
)
coef(fit_trg)
}
