% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.coxtrans.R
\name{predict.coxtrans}
\alias{predict.coxtrans}
\title{Prediction method for \code{coxtrans} objects.}
\usage{
\method{predict}{coxtrans}(
  object,
  newdata = NULL,
  newgroup = NULL,
  type = c("lp", "terms", "risk", "expected", "survival"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{coxtrans}.}

\item{newdata}{Optional new data for making predictions. If omitted,
predictions are made using the data used for fitting the model.}

\item{newgroup}{Optional new group for making predictions. If omitted,
predictions are made using the groups from the original data.}

\item{type}{The type of prediction to perform. Options include:
\describe{
\item{\code{"lp"}}{The linear predictor.}
\item{\code{"terms"}}{The components of the linear predictor.}
\item{\code{"risk"}}{The risk score \eqn{\exp(\text{lp})}.}
\item{\code{"expected"}}{The expected number of events, given the
covariates and follow-up time.}
\item{\code{"survival"}}{The survival probability, given the covariates
and follow-up time.}
}}

\item{...}{Additional arguments (not unused).}
}
\value{
A numeric vector of predictions.
}
\description{
Prediction method for \code{coxtrans} objects.
}
