% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ncvcox.R
\name{predict.ncvcox}
\alias{predict.ncvcox}
\title{Predict method for ncvcox objects}
\usage{
\method{predict}{ncvcox}(object, newdata, type = c("risk", "lp"), ...)
}
\arguments{
\item{object}{an object of class "ncvcox".}

\item{newdata}{a data frame in which to look for variables with which to
predict.}

\item{type}{the type of prediction to be carried out. The default is "risk".}

\item{...}{Unused.}
}
\value{
a vector of predicted values.
}
\description{
Predict method for ncvcox objects
}
