% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvcox.R
\name{predict.ncvcox}
\alias{predict.ncvcox}
\title{Prediction method for \code{ncvcox} objects.}
\usage{
\method{predict}{ncvcox}(object, newdata = NULL, newgroup = NULL, type = c("lp", "risk"), ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{newdata}{Optional new data for making predictions. If omitted,
predictions are made using the data used for fitting the model.}

\item{newgroup}{Optional new group for making predictions. If omitted,
predictions are made using the groups from the original data.}

\item{type}{The type of prediction to perform. Options include:
\describe{
\item{\code{"lp"}}{The linear predictor.}
\item{\code{"risk"}}{The risk score \eqn{\exp(\text{lp})}.}
}}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric vector of predictions.
}
\description{
Prediction method for \code{ncvcox} objects.
}
