% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2.R
\docType{data}
\name{sim2}
\alias{sim2}
\title{Simulated Group Survival Data: Heterogeneous Baseline Hazard Function and
Heterogeneous Effects}
\format{
A data frame with 1000 rows and 14 variables:
\describe{
\item{id}{Individual identifier, 1-1000.}
\item{group}{Group indicator, 1-10.}
\item{time}{Survival time.}
\item{status}{Status indicator, 0=censored, 1=event.}
\item{X1}{covariate 1.}
\item{X2}{covariate 2.}
\item{X3}{covariate 3.}
\item{X4}{covariate 4.}
\item{X5}{covariate 5.}
\item{X6}{covariate 6.}
\item{X7}{covariate 7.}
\item{X8}{covariate 8.}
\item{X9}{covariate 9.}
\item{X10}{covariate 10.}
}
}
\source{
\url{../articles/simulate-data.html}
}
\description{
A dataset containing 10 groups of survival data, where part of the effects
and baseline hazards are heterogeneous. Each group contains 100 individuals.
The survival time \eqn{T} of group \eqn{i} is generated according to the
following models:
\deqn{\lambda^{(i)}(t)=\lambda_{0}^{(i)}(t)\exp\left(x^\top(\beta+\eta^{(i)})
\right),} where \eqn{\lambda_{0}^{(i)}(t)} represents the baseline hazard of
group \eqn{i}, \eqn{\beta+\eta^{(i)}} represents the effects of group
\eqn{i}, and \eqn{x} represents the covariates. The covariates
\eqn{x=(x_{1},\ldots,x_{10})} are generated from a multivariate normal
distribution with mean 0 and covariance matrix \eqn{\Sigma=I_{10}}. The
baseline hazards are generated as follows:
\deqn{\lambda_{0}^{(i)}(t)=\left\{\begin{array}{ll}t^2,&i=1,3,\dots,9,\\
t,&i=2,4,\dots,10.\end{array}\right.}
while the effects \eqn{\eta^{(i)}} are generated as follows:
\deqn{\eta^{(i)}=\left\{\begin{array}{ll}(0,1,\ldots,0,1),&i=1,\dots,9,\\
(0,-1,\ldots,0,-1),&i=2,\dots,10.\end{array}\right.}
Additionally, \eqn{\beta=(1,0,\ldots,1,0)}. The maximum censoring time is
fixed at 3, and the censoring rate is approximately 20\%.
}
