% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsurv_tl.R
\name{simsurv_tl}
\alias{simsurv_tl}
\title{Simulate survival data of Multi-source Cox model}
\usage{
simsurv_tl(beta, eta, lambda, gamma, dist, maxt, n_samples, seed = 0)
}
\arguments{
\item{beta}{the common coefficients, a vector of length p,
where p is the number of features}

\item{eta}{the group-specific coefficients, a matrix of size p x K,
where K is the number of groups}

\item{lambda}{the baseline hazard's scale parameters, a vector of length K}

\item{gamma}{the baseline hazard's shape parameters, a vector of length K}

\item{dist}{the distribution of the baseline hazard, a string, either
"exponential","weibull", or "gompertz"}

\item{maxt}{the maximum time to simulate, a positive number}

\item{n_samples}{a vector of length K specifying the number of samples or a
single number specifying the total number of samples}

\item{seed}{the random seed, the default is 0}
}
\value{
a data frame with columns "id", "group", "X1", "X2", ..., "Xp",
"time", "status"
}
\description{
Simulate survival data of Multi-source Cox model
}
\examples{
beta <- c(1, 1)
eta <- matrix(c(0, 0, 1, 1), nrow = 2, ncol = 2)
lambda <- c(1, 2)
gamma <- c(2, 1)
dist <- c("gompertz", "weibull")
maxt <- 3
n_samples <- 100
df <- simsurv_tl(beta, eta, lambda, gamma, dist, maxt, n_samples)
head(df)
}
