% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ncvcox.R
\name{summary.ncvcox}
\alias{summary.ncvcox}
\title{Summary method for a \code{ncvcox} model}
\usage{
\method{summary}{ncvcox}(object, conf.int = 0.95, compressed = TRUE, ...)
}
\arguments{
\item{object}{the result of a \code{ncvcox} fit.}

\item{conf.int}{a numeric value between 0 and 1 indicating the confidence
level of the confidence interval. Default is 0.95.}

\item{compressed}{logical; if \code{TRUE}, the summary is compressed and
only includes the non-zero coefficients. Default is \code{TRUE}.}

\item{...}{Unused.}
}
\value{
A object of class \code{summary.ncvcox}, with the following
components:
\item{\code{n}, \code{nevent}}{number of observations and number of events,
respectively, in the fit}
\item{\code{logLik}}{the log partial likelihood at the final value}
\item{\code{BIC}}{the Bayesian Information Criterion at the final value}
\item{\code{coefficients}}{a matrix with one row for each coefficient, and
columns containing the coefficient, the hazard ratio exp(coef), standard
error, Wald statistic, and P value}
\item{conf.int}{a matrix with one row for each coefficient, containing the
confidence limits for exp(coef)}
}
\description{
Summary method for a \code{ncvcox} model
}
