% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtrans-package.R
\docType{package}
\name{survtrans-package}
\alias{survtrans}
\alias{survtrans-package}
\title{survtrans: Transfer Learning for Survival Analysis}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SignorinoY/survtrans}
  \item Report bugs at \url{https://github.com/SignorinoY/survtrans/issues}
}

}
\author{
\strong{Maintainer}: Ziyang Gong \email{gongziyang@smail.swufe.edu.cn}

}
\keyword{internal}
