% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtrans_control.R
\name{survtrans_control}
\alias{survtrans_control}
\title{Ancillary arguments for controlling survtrans fitting}
\usage{
survtrans_control(eps = 1e-06, maxit = 1000, patience = 10)
}
\arguments{
\item{eps}{the convergence criteria for the CCD algorithm. Default is 1e-6.}

\item{maxit}{the maximum number of iterations for the CCD algorithm.
Default is 1000.}

\item{patience}{the number of iterations without improvement before
stopping the CCD algorithm. Default is 10.}
}
\description{
Ancillary arguments for controlling survtrans fitting
}
