% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtrans_control.R
\name{survtrans_control}
\alias{survtrans_control}
\title{Ancillary arguments for controlling survtrans fitting}
\usage{
survtrans_control(
  abstol = 1e-04,
  reltol = 0.001,
  nthreads = 1L,
  maxit = 300,
  verbose = FALSE
)
}
\arguments{
\item{abstol}{the absolute tolerance for ADMM primal/dual residuals.
Default is 1e-4.}

\item{reltol}{the relative tolerance for ADMM primal/dual residuals.
Default is 1e-3.}

\item{nthreads}{the number of OpenMP threads used by a single model fit.
Default is 1.}

\item{maxit}{the maximum number of iterations for the proposed algorithm.
Default is 300.}

\item{verbose}{a logical value indicating whether to print messages
during the fitting process. Default is \code{FALSE}.}
}
\value{
A list with components \code{abstol}, \code{reltol},
\code{nthreads}, \code{maxit}, and \code{verbose}.
}
\description{
Ancillary arguments for controlling survtrans fitting
}
\keyword{internal}
