% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.ncvcox.R
\name{vcov.ncvcox}
\alias{vcov.ncvcox}
\title{Variance-covariance matrix for a \code{ncvcox} object.}
\usage{
\method{vcov}{ncvcox}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ncvcox}.}

\item{...}{Additional arguments (unused).}
}
\value{
A matrix representing the variance-covariance matrix of the
coefficients.
}
\description{
Variance-covariance matrix for a \code{ncvcox} object.
}
